🔥 remove CardWithTitleAndText
