--- nativeApp/src/screens/admin_views/Projects/EditProject/index.tsx
@@ -1,25 +1,0 @@
-import React, { FC } from 'react';
-// import styled from 'styled-components/native';
-import Page from '../../../../lib/ui/Page';
-
-/*
- * Types
- */
-type Props = {
-}
-
-/*
- * Styles
- */
-
-
-/*
- * Component
- */
-const EditProject: FC<Props> = () => (
-  <Page heading="Edit Project">
-    {}
-  </Page>
-);
-
-export default EditProject;
