+++ nativeApp/src/screens/admin_views/Projects/ProjectsPage/ProjectCard.tsx
@@ -1,9 +1,12 @@
 import React, { FC } from 'react';
+import styled from 'styled-components/native';
 
 import { MaterialIcons } from '@expo/vector-icons';
-import CardWithTitleAndDate from '../../../../lib/ui/CardWithTitleAndDate';
+import { withNavigation, NavigationInjectedProps } from 'react-navigation';
 import { ColoursEnum } from '../../../../lib/ui/colours';
-import { RemovalType } from '../../../../lib/ui/CardWithButtons';
+import CardWithButtons, { RemovalType } from '../../../../lib/ui/CardWithButtons';
+import { FontsEnum } from '../../../../lib/ui/typography';
+
 /*
  * Types
  */
@@ -17,18 +20,48 @@ type Props = {
 /*
  * Styles
  */
+const Heading2 = styled.TextInput`
+  marginLeft: 5;
+  fontSize: 25;
+  color: ${ColoursEnum.darkGrey};
+  fontFamily: ${FontsEnum.light};
+  letterSpacing: 1.2;
+`;
+
+const HeadingContainer = styled.View`
+  width: 100%;
+  flexDirection: row;
+  alignItems: flex-end;
+  marginBottom: 5;
+`;
+
+const Description = styled.Text`
+  color: ${ColoursEnum.darkGrey};
+  fontSize: 15;
+  paddingBottom: 5;
+  marginLeft: 4;
+`;
+
 /*
  * Component
  */
-const ProjectCard: FC<Props> = (props) => {
-  const iconColour = props.removalType === 'archive' //eslint-disable-line
+const ProjectCard: FC<NavigationInjectedProps & Props> = ({ date, title, removalType }) => {
+  const iconColour = removalType === 'archive' //eslint-disable-line
     ? ColoursEnum.mustard
     : ColoursEnum.darkGrey;
   return (
-    <CardWithTitleAndDate {...props} datePrefix="Created">
-      <MaterialIcons name="assignment" outline size={35} color={iconColour} />
-    </CardWithTitleAndDate>
+    <CardWithButtons
+      onPressOne={() => {}}
+      onPressTwo={() => {}}
+      removalType={removalType}
+    >
+      <HeadingContainer>
+        <MaterialIcons name="assignment" outline size={35} color={iconColour} />
+        <Heading2 selectTextOnFocus>{title}</Heading2>
+      </HeadingContainer>
+      <Description>{`Created: ${date}`}</Description>
+    </CardWithButtons>
   );
 };
 
-export default ProjectCard;
+export default withNavigation(ProjectCard);
