+++ nativeApp/src/screens/admin_views/Volunteers/VolunteersPage/VolunteerCard.tsx
@@ -1,9 +1,11 @@
 import React, { FC } from 'react';
+import styled from 'styled-components/native';
 
 import { MaterialIcons } from '@expo/vector-icons';
 import { withNavigation, NavigationInjectedProps } from 'react-navigation';
-import CardWithTitleAndDate from '../../../../lib/ui/CardWithTitleAndDate';
 import { ColoursEnum } from '../../../../lib/ui/colours';
+import CardWithButtons from '../../../../lib/ui/CardWithButtons';
+import { Heading2 as H2 } from '../../../../lib/ui/typography';
 
 /*
  * Types
@@ -17,20 +19,39 @@ type Props = {
 /*
  * Styles
  */
+const Heading2 = styled(H2)`
+  marginLeft: 5;
+`;
+
+const HeadingContainer = styled.View`
+  width: 100%;
+  flexDirection: row;
+  alignItems: flex-end;
+  marginBottom: 5;
+`;
+
+const Description = styled.Text`
+  color: ${ColoursEnum.darkGrey};
+  fontSize: 15;
+  paddingBottom: 5;
+  marginLeft: 4;
+`;
 
 /*
  * Component
  */
-const VolunteerCard: FC<NavigationInjectedProps & Props> = ({ navigation, ...rest }) => (
-  <CardWithTitleAndDate
+const VolunteerCard: FC<NavigationInjectedProps & Props> = ({ navigation, date, title }) => (
+  <CardWithButtons
     onPressOne={() => navigation.navigate('AdminEditVolunteer')}
     onPressTwo={() => {}}
-    {...rest}
-    datePrefix="Joined"
     removalType="delete"
   >
-    <MaterialIcons name="person-outline" outline size={35} color={ColoursEnum.mustard} />
-  </CardWithTitleAndDate>
+    <HeadingContainer>
+      <MaterialIcons name="person-outline" outline size={35} color={ColoursEnum.mustard} />
+      <Heading2>{title}</Heading2>
+    </HeadingContainer>
+    <Description>{`Joined: ${date}`}</Description>
+  </CardWithButtons>
 );
 
 export default withNavigation(VolunteerCard);
